refactor: delete dir index.ts
