+++ .hygen/feature/new/.eslintrc.json.ejs.t
@@ -5,6 +5,6 @@ after: "// HYGEN_INJECT"
 ---
         {
           "module": "features/<%= feature_name %>/**/*",
-          "allowReferenceFrom": ["src/page-components/**/*"],
-          "allowSameModule": false
+          "allowReferenceFrom": ["src/page-components/**/*", "src/stores/**/*", "src/providers/**/*"],
+          "allowSameModule": true
         },
