--- src/features/albums/api/index.ts
@@ -1,4 +1,0 @@
-/**
- * @package
- */
-export { useGetAlbums } from './get-albums';
