--- src/features/albums/components/index.ts
@@ -1,1 +1,0 @@
-export { Albums } from './Albums';
