--- src/features/albums/hooks/index.ts
@@ -1,4 +1,0 @@
-/**
- * @package
- */
-export { useFetchAlbums } from './useFetchAlbums';
