+++ src/features/albums/index.ts
@@ -1,0 +1,1 @@
+export { Albums } from './components/Albums';
