--- src/features/albums/types/index.ts
@@ -1,9 +1,0 @@
-/**
- * @package
- */
-export type { Album, Albums } from './albums-type';
-
-/**
- * @package
- */
-export { AlbumsEntity } from './albums-type';
