--- src/features/posts/apis/index.ts
@@ -1,4 +1,0 @@
-/**
- * @package
- */
-export { useGetPosts } from './get-post';
