--- src/features/posts/components/index.ts
@@ -1,1 +1,0 @@
-export { Posts } from './Posts';
