--- src/features/posts/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useFetchPosts } from './useFetchPosts';
