--- src/features/posts/types/index.ts
@@ -1,9 +1,0 @@
-/**
- * @package
- */
-export type { Post, Posts } from './posts-type';
-
-/**
- * @package
- */
-export { PostsEntity } from './posts-type';
