+++ src/features/posts/types/posts-type.ts
@@ -8,19 +8,10 @@ const zPostScheme = z.object({
 });
 const zPostsScheme = z.array(zPostScheme);
 
-/**
- * @package
- */
 export type Post = z.infer<typeof zPostScheme>;
 
-/**
- * @package
- */
 export type Posts = z.infer<typeof zPostsScheme>;
 
-/**
- * @package
- */
 export class PostsEntity {
   readonly body!: string;
   readonly id!: number;
